# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Gym CRM API
  description: API documentation for Gym Customer Relationship Management system
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local Development Server

paths:
  /users/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: Successful login
        '401':
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{username}/password:
    put:
      summary: Change user password
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        '200':
          description: Password changed successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '422':
          description: Invalid new password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{username}/toggle:
    patch:
      summary: Toggle user activation status
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User activation status updated
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainees:
    post:
      summary: Create new trainee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTraineeRequest"
      responses:
        '201':
          description: Trainee created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsResponse"
        '422':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainees/{username}:
    get:
      summary: Get trainee profile
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trainee found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraineeResponse"
        '404':
          description: Trainee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update trainee profile
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTraineeRequest"
      responses:
        '200':
          description: Trainee updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraineeResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Trainee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '422':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete trainee
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Trainee deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainees/{username}/favorite-trainers:
    put:
      summary: Update trainee's favorite trainers
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFavoriteTrainersRequest"
      responses:
        '200':
          description: Trainee's favorite trainers updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShortTrainerResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Trainee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '422':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainees/{username}/trainings:
    get:
      summary: Find all trainee's trainings
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: trainer
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Find trainee's training successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Trainee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainers:
    get:
      summary: Get all trainers not assigned on trainee
      security:
        - basicAuth: []
      parameters:
        - name: unassigned
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: All not assigned trainers found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShortTraineeResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Trainee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create trainer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainerRequest"
      responses:
        '200':
          description: Trainer created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsResponse"
        '422':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainers/{username}:
    get:
      summary: Get trainer profile
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trainer found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainerResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Trainer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update trainer profile
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrainerRequest"
      responses:
        '200':
          description: Trainer updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainerResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Trainer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '422':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainers/{username}/trainings:
    get:
      summary: Get all trainer's trainings
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: trainee
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Trainer's trainings found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Trainer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainings:
    post:
      summary: Create training
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingRequest"
      responses:
        '201':
          description: Training created successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '422':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /training-types:
    get:
      summary: Get all training types
      security:
        - basicAuth: []
      responses:
        '200':
          description: All training types found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingTypeResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      required:
        - oldPassword
        - newPassword
    CreateTraineeRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
      required:
        - firstName
        - lastName
    UpdateTraineeRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
        isActive:
          type: boolean
      required:
        - firstName
        - lastName
        - isActive
    UpdateFavoriteTrainersRequest:
      type: object
      properties:
        trainers:
          type: array
          items:
            type: string
      required:
        - trainers
    CreateTrainerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        specialization:
          type: integer
      required:
        - firstName
        - lastName
        - specialization
    UpdateTrainerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        specialization:
          type: integer
        isActive:
          type: boolean
      required:
        - firstName
        - lastName
        - specialization
        - isActive
    CreateTrainingRequest:
      type: object
      properties:
        trainee:
          type: string
        trainer:
          type: string
        name:
          type: string
        date:
          type: string
          format: date
        duration:
          type: integer
      required:
        - trainee
        - trainer
        - name
        - date
        - duration
    CredentialsResponse:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    TraineeResponse:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        isActive:
          type: boolean
        favoriteTrainers:
          type: array
          items:
            $ref: "#/components/schemas/ShortTrainerResponse"
    TrainerResponse:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        specialization:
          type: integer
        isActive:
          type: boolean
        favoriteBy:
          type: array
          items:
            $ref: "#/components/schemas/ShortTraineeResponse"
    ShortTraineeResponse:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    ShortTrainerResponse:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        specialization:
          type: integer
    TrainingResponse:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
        duration:
          type: integer
        trainee:
          type: string
        trainer:
          type: string
    TrainingTypeResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        title:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string